requirements.txt
antlr4-python3-runtime==4.9.3
anyio==4.11.0
asttokens==3.0.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.3
contourpy==1.3.0
cvzone==1.6.1
cycler==0.12.1
debugpy==1.8.15
decorator==5.2.1
executing==2.2.0
filelock==3.16.1
filterpy==1.4.5
fonttools==4.54.1
fsspec==2024.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hydra-core==1.3.2
idna==3.10
imageio==2.35.1
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==6.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
Pygments==2.19.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-telegram-bot==22.5
pytz==2024.2
pywin32==311
PyYAML==6.0.2
pyzmq==27.0.1
requests==2.32.3
scikit-image==0.24.0
scipy==1.14.1
seaborn==0.13.2
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.13.3
tifffile==2024.9.20
torch==2.4.1
torchvision==0.19.1
tornado==6.5.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.2.101
ultralytics-thop==2.0.8
urllib3==2.2.3
wcwidth==0.2.13
